requirements.txt
# Core GUI framework
PySide6>=6.5.0

# Video processing and extraction
yt-dlp>=2024.1.0
python-vlc>=3.0.0

# YouTube transcript support
youtube-transcript-api>=1.0.0

# HTTP requests (usually included with Python)
requests>=2.28.0

# Development and testing dependencies (optional)
# black>=22.0.0  # Code formatting
# pytest>=7.0.0  # Testing framework
# ruff>=0.1.0    # Linting